Add files via upload
Just experimenting, thus so many commented lines of code, since I had trouble getting it to properly count the neighbor cells. But I finally figured it out through the use of the "%" modulo operator. Will clean up and create the "full" Conway's game of life in the next several changes.